Rename math to math_functions
