+++ src/math_functions.h
