+++ src/rand.h
@@ -103,6 +103,4 @@ inline int iRand_round(int max)
     return i/2;
 }
 
-extern int mathSign(int a);
-
 #endif // RAND_H
